move build command to own file
