+++ pkg/commands/main.go
@@ -1,7 +1,6 @@
 package commands
 
 import (
-	"flag"
 	"fmt"
 	"os"
 
@@ -16,24 +15,7 @@ var stderr = os.Stderr
 // Main is the root command of the whole CLI. It is given the name "go" so that
 // when this CLI is incorporated into a parent CLI, the commands within will be
 // rooted at "go". E.g. "go-build", "go-build primary", "go-build verification".
-var Main = cli.RootCommand("go-build", "go build and related functions",
-	Build, Verify, Config)
-
-type buildOpts struct {
-	buildish
-	verification bool
-	clean        bool
-}
-
-func (opts *buildOpts) Flags(fs *flag.FlagSet) {
-	opts.buildish.Flags(fs)
-	fs.BoolVar(&opts.verification, "verification", false, "configure build as a verification build")
-	fs.BoolVar(&opts.clean, "clean", false, "fail unless worktree is clean")
-}
-
-var Build = cli.LeafCommand("build", "run a build", func(opts *buildOpts) error {
-	return opts.runBuild("Running build", opts.verification, build.WithCleanOnly(opts.clean))
-})
+var Main = cli.RootCommand("go-build", "go build and related functions", Build, Verify, Config)
 
 var Describe = cli.RootCommand("describe", "describe things", DescribeBuildEnv)
